Make synchronous transfer APIs robust against signal interruption
libusb_control_transfer and libusb_bulk_transfer are designed to be
synchronous such that control is not returned until the transfer
definitively succeeds or fails.  That assumption is violated if a signal
interrupts these functions because there is no way for the application
to continue waiting for the transfer without resubmitting it.  This
patch changes these synchronous APIs so they do not abort in the case of
a signal interruption.

Signed-off-by: David Moore <dcm@acm.org>