linux: Stop the kernel from re-attaching in kernel drivers after a reset
When an interface is bound to the usbfs driver (iow claimed), the kernel will
unbind it, and then after the reset do a device_attach on the interface,
which will bind the default in kernel driver to the interface.

So if an app has detached the in kernel driver, and claimed the interface
and then does a libusb_reset_device. Things end up with the interface no
longer being bound to the usbfs driver (so no longer claimed) and instead it
is bound to the in kernel driver (iow the in kernel driver is re-attached).

We can stop this from happening by releasing all claimed interfaces before
the reset, as the kernel will not do the device attach after reset, if no
driver was bound to the interface before the reset.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>