Add libusb_has_capability() API function
Since it is currently not planned to change the filename of the libusb
DLL file for Windows systems between libusb-1.0 versions it is important
to have this API available starting with the first release that supports
Windows.

Currently there exists only one capability; LIBUSB_CAN_GET_DEVICE_SPEED,
which will always be supported when libusb_has_capability() is available.