WinCE: Fix device reference leak which caused crash on libusb_exit().
The Windows CE device allocation code has always had a bug where it would
leak references to devices when they are allocated. This commit removes the
reference leak.

This leak was highlighted by the new hotplug code which now triggers a NULL
pointer dereference if not all devices are unreferenced before libusb_exit
is called.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>