Fixed crash in Windows due to modified usbi_os_backend structre
Since the initialization of the windows_backend does not use named initializers, the addition of the open_fd function pointer to the usbi_os_backend structure caused the initializers in windows_backend to be offset resulting in most of the function pointers to be incorrect. This has now been fixed by adding in an initializer to set the open_fd function pointer to NULL for the windows_backend.