Remove unused member add_iso_packet_size from struct usbi_os_backend
The member add_iso_packet_size of struct usbi_os_backend is set to 0 in
all backends and thus has no actual use. This has been the case since
commit ad6e2b71 ("Linux: fire multiple URBs at once for split
transfers").

Also, the comment above the member actually states the fact that it is
unused and could potentially be removed, so remove it.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>