Use calloc to allocate and zero out buffers where appropriate
Use calloc() instead of malloc() + memset(..., 0, ...) and get rid of
some now unnecessary variables.

Since in these two cases an array of elements is allocated, calloc()
make more sense semantically as well and it checks for overflows in the
calculation of the buffer size (though that shouldn't ever be the case
here).

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>