linux_usbfs: Set reap action to CANCELLED only when successful
Prior to this commit, a call to libusb_cancel_transfer() would set
the reap action of the transfer to CANCELLED regardless of whether
the cancellation was successful or not. This can cause the transfer
status to incorrectly report LIBUSB_TRANSFER_CANCELLED when in some
cases this is not the case. This commit adds a check for a successful
cancellation before setting the reap action as such.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>