linux: Correctly return device configuration when using usbfs
libusb_get_configuration() has been broken when using a Linux kernel
that requires the use of usbfs to obtain the current configuration
value. This commit resolves this issue and also modifies the
usbfs_get_active_config() function to update the cached active
configuration value itself instead of requiring the callers to do it.

Closes #137

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>