fix timerfd initialization
sys/timerfd.h defines TFD_NONBLOCK as 0x800 but in kernel TFD_NONBLOCK
is an alias for O_NONBLOCK which is defined in arch-specific fcntl.h.
While it's still 0x800 for most of archs but for mips it's 0x80. So
timerfd_create(..., TFD_NONBLOCK) returns -EINVAL because of that. Fix
this by using O_NONBLOCK instead.

Signed-off-by: Alexander Gordeev <lasaine@lvk.cs.msu.su>