linux_usbfs: Disable hotplug events and scanning on Android
On Android, the platform API should be used to scan for and open devices
and pass file descriptors to libusb.  Newer devices (Android 5+) even
prohibit listening for hotplug events, resulting in libusb failing to
initialize without this patch.

Note that this patch effectively renders libusb useless on older devices
that do not have USB support in the platform API (anything before
Android 5).

Signed-off-by: Vianney le Clément de Saint-Marcq <code@quartic.eu>