Windows: Provide C99 compliant implementations of (v)snprintf()
The Microsoft implementations of _snprintf() and vsnprintf() do not
guarantee that the output buffer is NULL-terminated, and they do not
return the number of bytes that *would* have been written if truncation
were to occur. This commit adds implementations that do just that.

Note that VS2015 and above provide C99 compliant versions of these
functions, so those will be used when available.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>