Replaced low-level mach API with pthread API
Found this because clang was warning that the return value was being converted from unsigned to signed.

The need for a mach API surprised me, and I found this alternative formulation in a Chromium bug:

https://groups.google.com/forum/#!topic/native-client-reviews/z---hlHbQNc

which describes: "mach_thread_self() needs to be balanced by mach_port_deallocate(), which is two system calls. pthread_mach_thread_np(pthread_self()) is two libc function calls and no system calls, because pthread caches the Mach port."

Fixed the conversion warning too, with a cast.